Test that the controller creates a git source
The idea of an assemblage is that it will create all the sources and
kustomizations implied by the definition. This adds a test which
checks the very least it should do, which is to create a source.

The test fails, at this point -- the controller is yet to be written.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>