Add minimal fields to RemoteAssemblage
This commit gives RemoteAssemblage some shape: it needs a reference to
another cluster, and a specification for the assemblage to create
there.

On the matter of how to specify the downstream assemblage, there are
many directions for the design to go (using a template for the
assemblage, or otherwise indirecting, ...). The simplest way to get
something workable is to inline the assemblage spec. When other
considerations come to be addressed, like parameterising assemblages,
this will surely change.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>