Set up environment for remote syncing
Here I've borrowed some test code from github.com/squaremo/starling,
which sets up a downstream test env, and makes a record for it in the
main test env. This can then be used to check whether a controller
effects changes downstream (and reflects status changes upstream).

The biggest trouble was finding compatible versions of
controller-runtime and cluster-api. Released versions of cluster-api
depend on an ancient controller-runtime (v0.5). The only configuration
which compiled was to get cluster-api from master branch, and the
v0.9.0-alpha.0 release of controller-runtime.

The actual test is a placeholder and not expected to pass yet.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>