Maintain summary of module use in module.status
A module is assigned to a set of clusters, and for each cluster will
be in a particular state -- either updated but not synced,
successfully synced, or failed to sync.

This commit adds a summary of the states to a module's status, with a
view towards doing incremental rollouts of new module versions.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>