Change how module status summary is calculated
Usually you want the status to be internally consistent; e.g., in the
case of a Module, for the summary to give counts related to the source
version reported.

To make this the case in module_controller.go, the summary must be
calculated while the spec is applied. If a particular assemblage gets
changed, then it's necessarily updating (at least until the next time
you look), and if it didn't, then it might still be updating, or have
succeeded or failed.