Run assemblage tests in workflow
This adds a GitHub Actions workflow that will run the tests and docker
image build for assemblage/.

There is some extra setup needed, arising mainly from being a private
Git repo: the build needs an SSH key for `go mod` to access the
repository; the private key goes in project secrets, and the public
key is added as a deploy key with read access. An ssh-agent is run so
that the docker-build can take advantage of the key too. This
replicates steps taken for the docker build to work in a local
environment, in fdcb81802b3f4f0b40c6a6dcd70f841f180f3a6b.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>