Prepare bindings types for module layer
In the module layer, there are two kind of bindings, those evaluated
in the control plane, and those evaluated in the target cluster. For
BootstrapModule, only those evaluated in the control plane are
allowed; so, to be able to use the Sync type there, it cannot have the
bindings included. Howeve, they _can_ go in NamedSync, since that is
exactly the type used when bindings are included in the sync -- that
is, in the assemblage types.

While I'm here: using a *string for the value is awkward, because you
can't take the pointer of a string literal. so, move the value into a
struct, with which you can do that. This makes the syntax a little
weird, though:

```
bindings:
- name: APP_NAME
  value:
    string: foo
```

I will have to come up with another formulation.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>