Add controlPlaneBindings to BootstrapModule
BootstrapModule objects specify a sync to apply directly to the remote
cluster. Since there is no controller doing further work in the remote
cluster, there's no bindings evaluated there. (Technically, that could
be done remotely too, but it is simpler for implementation and
comprehension to disallow them).

This commit adds controlPlaneBindings to the BootstrapModule type, for
supplying bindings to be evaluated in the control plane.

It's desirable to let people mention previous (or just other) binding
names when defining a binding. For example,

```yaml
bindings:
- name: HOST
  value: service.cluster.local
- name: PORT
  value: 3030
- name: HOSTPORT
  value: "$(HOST):$(PORT)"
```

My first punt at resolving bindings that might mention other bindings
was to supply the map of memoised values, with the idea being that if
the bindings are in dependency order, the map will have a value for
any bindings needed -- in the example above, by the time it gets to
HOSTPORT, it has evaluated HOST and PORT.

The problem with that is bindings are evaluated by need; that is, when
_mentioned_. So if HOSTPORT is used in the sync spec, but HOST or PORT
aren't, then it will not have a value for those latter when it comes
to evaluate the former.

The solution is to recurse on binding resolution. To detect circular
definitions, a stack keeps track of which bindings were being resolved
before seeing attempting to resolve this bindings.

This has the property that the bindings can be defined in any order,
but I don't want people to rely on that. It's just easier to implement
that way!

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>