Support recursive bindings in bootstrap modules
This commit rewrites the binding resolution used in the
BootstrapModule controller, so that it will properly resolve bindings
that mention other bindings.

I have done this a little different to the original implementation in
assemblage_controller.go, to see if it's neater. A third
implementation (for Modules), with slightly different requirements
again, will help me see how best to factor things.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>