Implement bindings for Module objects
Module differs again from Assemblage and BootstrapModule, since it
does not expand directly to Kustomizations, and can have bindings that
evaluate in the target cluster as well as bindings that evaluate in
the control plane.

The central trickiness here is that bindings for the control plane
must be evaluated before the module is "sent" to each cluster (i.e., a
remote assemblage is updated).

I have done the naive thing for now, and evaluated each control plane
binding to get a value, which is sent to the target cluster. An
obvious improvement is to expand (but not evaluate) the target cluster
sync, and thereby determine which control plane bindings are needed,
and their values.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>