Revisit syntax for "value" binding
I want a syntax for value bindings (where the referent is a literal
string that gets expanded) to be as simple as possible. But I also
want it to work with the usual way of doing union types, which is to
make each member of the union a pointer type (why? the zero value is
`nil`, so for ground types it is possible to distinguish between "not
supplied" and "supplied but the zero value").

The first take was:

```go
type BindingSource struct {
  ObjectFieldRef *ObjectFieldSelector `json:"objectFieldRef,omitempty"`
  Value: *string `json:"value,omitempty"`
}
```

which results in the desired schema, but is a pain to use in Go
because you can't take the pointer of a literal string.

Second take:

```go
type BindingSource struct {
  ObjectFieldRef *ObjectFieldSelector `json:"objectFieldRef,omitempty"`
  Value *Value `json:",omitempty"`
}

type Value struct {
  String: string `json:"string"`
}
```

This avoids awkward pointering in Go; but doesn't have the compact
YAML representation, since you now have to use both the value and
string fields.

This is the third take, which is:

```go
type BindingSource struct {
  ObjectFieldRef *ObjectFieldSelector `json:"objectFieldRef,omitempty"`
  *StringValue `json:",omitempty"`
}

type Value struct {
  Value: string `json:"value"`
}
```

Embedding the struct makes the schema generator inline the `Value`
field, giving the desired YAML syntax; but it's still a pointer-valued
field.