Temporarily use fleeet/pkg module from filesystem
To make it easier to develop changes in github.com/squaremo/fleeet/pkg
along with its use in the controllers, I've put a replace directive in
go.mod. This will break CI for now, since that tests the image build,
and the image build (correctly!) doesn't include pkg/ in its build
context.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>