RFC on rollout automation
This adds a "work in progress" RFC on changing the layering to make
rollouts possible. It's an RFC, rather than a design, because it
proposes changes to the design -- if I can get it into a good shape
(including, possibly, a reference implementation), then I'll update
the design docs as a follow-up.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>