Remove Partition type in favour of direct relation
The partition type was the locus for the module->cluster assignment,
with the idea being that a rollout controller can change which
clusters belong to which partitions to effect a rollout.

However, it's not necessary to have an intermediate object for this
relation -- clusters can simply be owned by modules directly, without
losing generality. Since partitions don't pay their way, I've removed
them from the design.