Rework how assignment is represented
Using owner references will interact with Kubernetes' garbage
collection. Even if it's possible to work around that with e.g.,
finalisers, it seems like a bad idea to subvert GC. It also doesn't
help with atomicity, for which other workarounds are needed.

Instead, use the object that already represents the assignment of
modules to a cluster -- the RemoteAssemblage -- and change its meaning
to suit. More shuffling may be necessary to make BootstrapModules fit
too.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>