Remove the extra layer in rollout RFC
After realising that assemblages represent the assignment of modules
to clusters -- thus no need for another object -- a further
realisation is that building a Rollout layer on top of modules makes
modules essentially passive objects. Another way to look at it is that
Modules may as well be the locus for rollout logic, since otherwise
they do not have a job to do.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>