Add ProxyAssemblage type
    kubebuilder create api --kind ProxyAsssemblage ...

One change needed for rollouts is to have an "assemblage" object to be
the place all the assigned bootstrap modules (syncs which get applied
remotely from the control plane) get recorded for each cluster. A good
name for that would be `RemoteAssemblage` (e.g., assemblage which is
applied remotely) -- but that name is already used for the one that
proxies to a remote cluster, which would be better-named
`ProxyAssemblage`. To create myself some bother, I'm adding the new
type and the rename at the same time.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>