Copy implementation for remote to proxy
As part of the rename of RemoteAssemblage to ProxyAssemblage, this
commit copies the implementation of the remote assemblage controller
to the proxy assemblage controller, ports the tests for the former to
the latter, and likewise swaps the type used in module_controller.go.

The implementation for (repurposed) remote assemblages will be new,
and need new tests.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>