Expand bootstrap modules via remote assemblages
At present, bootstrap modules get expanded directly into Flux
primitives (one GitRepository, and a Kustomization per
cluster). Incremental rollouts could be implemented within this
scheme; however, it would not help third party integrations much since
they'd have to implement all of it themselves. Instead, I'm
repurposing the RemoteAssemblage type -- the behaviour of which was
moved to the better-named ProxyAssemblage -- to represent a set of
syncs to be applied remotely to a cluster.

In this commit:

 - change the definition of the RemoteAssemblage type so it contains
   syncs to apply remotely, rather than syncs to proxy (those were
   moved to ProxyAssemblage)

 - move the "expand to Flux primitives" code from the bootstrap module
   controller to the remote assemblage controller

 - implement the construction of remote assemblages in the bootstrap
   module controller

 - fix up type names (remote v proxy) in RBAC anontations and comments
 - adapt test code to the above changes

The aim is to eventually put the commonly useful bits -- expansion to
Flux primitives, and binding evaluation -- in the assemblage code, and
the rollout logic in the module code. An integration or extension can
then replace the module part by building on the assemblage part.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>