Avoid a sync vs cluster ready race
When you create a cluster with Cluster API, it will typically take
some minutes between the Cluster object (and its companions) appearing
and the provider creating a secret containing a kubeconfig for the
newly-provisioned cluster. If you have a bootstrap module already, the
Flux primitives targeting the new cluster will fail in that period,
possibly leading to a long back-off.

There's at least a couple of places this can be mitigated:

 - only create a RemoteAssemblage when a cluster is marked ready;
 - only create Flux primitives when the required secret exists;

Here, I've implemented the latter, since it is at a lower level and
therefore will apply more broadly.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>