Factor out Kustomize ready-state func
The assemblage controller has a func for calculating the ready state
of the Kustomization objects it created. This is useful for the remote
assemblage controller too, which is in a different layer (module
rather than assemblage) -- so, recreate it in pkg/sync to be available
to both layers.

I changed the argument type; it seems like a calculation specific to
the Flux Kustomize type. But if I need it for other things, perhaps it
can be generalised, or just replaced with kstatus.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>