Watch owned objects to react to OOB changes
It's expected (and useful) behaviour to recreate a controller-created
object that is deleted out of band (e.g., by a pesky human). To make
it work this way, it's usually only necessary for the controller to
watch the objects in question, since a deletion will then trigger a
reconciliation of the owner object.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>