Poll for assemblage updates in the downstream
The ProxyAssemblage controller does not have the opportunity to be
notified of changes in a downstream cluster (or at least, this would
be tricky to arrange, especially at scale), so it won't in general see
when its Assemblage counterpart changes status.

A dumb but effective way to make sure updates are seen is to simply
requeue each ProxyAssemblage to be re-examined after a small
interval. This will be OK at moderate scale (I would guess O(100)
clusters), since more worker threads can be configured to cope with
polling. Larger scales may need a less brute force approach to be
figured out.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>